# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sample API",
    "description": "API description in Markdown.",
    "contact": {
      "email": "joaotavares@dev.com",
      "name": "Joao Tavares"
    },
    "version": "1.0.0"
  },
  "host": "api.example.com",
  "basePath": "/v1",
  "schemes": ["https"],
  "paths": {
...
    "/user": {
      "post": {
        "summary": "Create user",
        "tags": ["User"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              },
              "example": {
                "name": "Joao Tavares",
                "email": "joaotavares@dev.com",
                "birthDate": "2018-12-10T13:49:51.141Z",
                "password": "pa$$w0rd",
                "avatar": "avatar.com.br"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                },
                "example": {
                  "data": {
                    "id": "cf58b454-cd06-43df-87ed-16bc3a5f9e0e",
                    "name": "Joao Admin",
                    "email": "joao7711@gmail.com",
                    "avatar": "avatar.com.br",
                    "role": "CUSTOMER",
                    "birthDate": "2018-12-10T13:49:51.141Z",
                    "password": "$2a$08$CTS6K1qKW2HMclFK9fzUb.lqtaCyUwb43r3EzyIMTVT.gKvpAYhG2",
                    "createdAt": "2023-05-30T18:37:14.977Z",
                    "updatedAt": "2023-05-30T18:37:14.977Z",
                    "externalId": null
                  },
                  "error": null,
                  "statusCode": 201
                }
              }
            }
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "tags": ["User"],
        "summary": "Get User",
        "description": "Get user by ID",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token",
            "required": true,
            "type": "string",
            "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6ImFjOGQwNGE0LTRiOTYtNGQyNS04MjFmLWZiMWU3MDkyMzYwOCIsImlhdCI6MTY4NTUzMTYzOCwiZXhwIjoxNjg1NjE4MDM4fQ.kfHxR-M4HBznkkkRuveUkFD8HRTbbJkdstAblsz6oO5"
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "type": "string",
            "example": "cf58b454-cd06-43df-87ed-16bc3a5f9e1u"
          }
        ],
        "responses": {
          "201": {
            "description": "User found!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                },
                "example": {
                  "data": {
                    "id": "cf58b454-cd06-43df-87ed-16bc3a5f9e0e",
                    "name": "Joao Admin",
                    "email": "joao7711@gmail.com",
                    "avatar": "avatar.com.br",
                    "role": "CUSTOMER",
                    "birthDate": "2018-12-10T13:49:51.141Z",
                    "password": "$2a$08$CTS6K1qKW2HMclFK9fzUb.lqtaCyUwb43r3EzyIMTVT.gKvpAYhG2",
                    "createdAt": "2023-05-30T18:37:14.977Z",
                    "updatedAt": "2023-05-30T18:37:14.977Z",
                    "externalId": null
                  },
                  "error": null,
                  "statusCode": 201
                }
              }
            }
          }
        }
      }
    },
    "/user/image": {
      "put": {
        "tags": ["User"],
        "summary": "PUT image",
        "description": "Update user avatar",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token",
            "required": true,
            "type": "string",
            "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6ImFjOGQwNGE0LTRiOTYtNGQyNS04MjFmLWZiMWU3MDkyMzYwOCIsImlhdCI6MTY4NTUzMTYzOCwiZXhwIjoxNjg1NjE4MDM4fQ.kfHxR-M4HBznkkkRuveUkFD8HRTbbJkdstAblsz6oO5"
          },
          {
            "name": "image",
            "in": "formData",
            "description": "Image file",
            "required": "true",
            "type": "file"
          }
        ],
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object"
                    },
                    "error": {
                      "type": ["null", "string"]
                    },
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    }
                  },
                  "example": {
                    "data": {},
                    "error": null,
                    "statusCode": 200
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date-time"
          },
          "password": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          }
        },
        "required": ["name", "email", "birthDate", "password"]
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "avatar": {
                "type": "string",
                "nullable": true
              },
              "role": {
                "type": "string",
                "enum": ["CUSTOMER", "ADMIN"]
              },
              "birthDate": {
                "type": "string",
                "format": "date-time"
              },
              "password": {
                "type": "string"
              },
              "createdAt": {
                "type": "string",
                "format": "date-time"
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time"
              },
              "externalId": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "id",
              "name",
              "email",
              "role",
              "birthDate",
              "password",
              "createdAt",
              "updatedAt"
            ]
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": ["data", "statusCode"]
      },
...
    }
  }
}